Implement a Dart Gemini client and samples
Design follows the general approach of the generative AI SDKs in other
languages.

A `GenerativeModel` class is initialized with an API key and a
chosen model. The model communicates with the remote API using REST
requests. It implements the `generateContent`, `generateContentStream`,
`countTokens`, `embedContent` APIs. Outer level "request" objects are
not exposed - the fields of those objects are the arguments to the model
methods. This reduces noise at the callsites by avoiding a nominal
constructor with a long name.

Implement a `text` getter on the response to generate content requests
which simplifies extracting a String from the candidate and part lists
in the common case of wanting the starting text from the first
candidate.

A `ChatSession` can be created from a model by calling an extension
method. The chat session holds chat history

Data models, translation to json and parsing from json, are hand-written
against the proto definitions for the API. We may consider
autogenerating these from some source in the future.

The `Content` data model is the primary argument for most of the service
methods, and has a very general design which specifies a "role" and
allows the content to come in multiple parts. The typical case for
arguments to the generative model methods will be to use the "user" role
and a single part. Add a `Content.text`, `Content.data`, and
`Content.multi` helpers for creating content with a user role.

Other repo changes:
- Add Dart related local files to `.gitignore.`
- Update README with dart code samples and early usage instructions.