Normalize model names
Strip a leading `models/` since the model names are often used with that
prefix. Store the field without the prefix so it can be used as a URI
path segment and no URI parts are constructed with string operations.

Change a call to `setUp` to define a method called from each test to
allow one case where the model name is overidden. Add a single test that
validates the  URI does not have a double `models/` segment.