Remove incorrect error handling in sample
The usage of `ParallelWaitError.errors` was incorrect - it is not an
iterable.
Remove the error handling, since it is not the interesting part of the
sample and its more noise than is useful.

Inline the variable declaration in both places where Record.wait is
used. In both cases the `late` was already unnecessary.