Add parsing for unary server errors (#36)
When parsing to a `GenerateContentResponse` fails with a format
exception, it may be due to the server responding with an "error"
response.

Add specific exception types for `InvalidApiKey` and
`UnsupportedUserLocation` since these errors may need specific handling
in a client. For now all other errors are a `ServerException`, but we
can add other types if it becomes useful to match more specific
patterns.

Add tests with scraped server error responses parsed to each error type.

Export the errors from the package public API.

Add a doc comment to the `GenerativeAIException` class from
https://github.com/google/generative-ai-dart/pull/28/files#diff-422b60b8e6460903ebe4ea03bef1a87c07ba50229d3f46bf436dc8ddd905d050

Fix a typo in the exception `toString()`.