Cleanup in sample code (#49)
Remove incorrect usage of `ParallelWaitError.errors` - it is not an
iterable. The error handling is not the interesting part of the sample
and it's more noise than is useful, so remove it. Inline the variable
declaration in both places where Record.wait is used. In both cases the
`late` was already unnecessary.

Apply fixes from https://github.com/google/generative-ai-dart/pull/28
- Write errors to `stederr`.
- Reference `message` var instead of repeat the string.
- Fix usage of Uri path to resolve resource files.