Improve chat implementation (#50)
Taken from https://github.com/google/generative-ai-dart/pull/28

Replace `package:pool` with a small `Mutex` implementation. We don't
need the full capabilities of `Pool`, so remove the dependency.

Improve doc comments - rephrase, add missing docs, and improve
formatting.

Make `sendMessageStream` a synchronous method with an inner `async*`
function literal. This allows taking the lock synchronously instead of
waiting for the returned stream to have a listener. Add a TODO about
eagerly listening and buffering the response so that it can be added to
history as it arrives.