Tweak the code, comments and utilities.
Nothing here should change behavior or API, only
implementation and presentation.

Introduced the `Mutex` class because `Pool` seemed
overqualified/overcomplicated for the job of being
a single-lock mutex.

Added a `JsonConvertible` interface, just to have
an extension method on `Iterable<JsonConvertible>`
for doing `[for (var v in iterable) v.toJson()]`.
Happened too often, so a good candidate for a
helper function.

Every change here is a suggestion, none of it is
essential. Feel free to ignore any parts that
seem unnecessary.
(The task-URI cache is probably unnecessary.)