Make `sendRequestStream` eager and caching.
The `sendRequestStream` function now sends the request eagerly,
as soon as prior requests are done, whether someone is listening
to the returned stream or not.
The stream will cache responses until listened to.

This should ensure that nobody listening on the stream won't
affect the ordering of requests, or block later requests.