Make `_aggregate` not copy single `TextPart`s
The `_aggregate` function combines adjacent `TextPart`s
in a sequence of lists of parts.
It was doing this by adding the text content of text parts
to a `StringBuffer`, then creating a new text part from the
result when no more text parts follows.
This change reuses the original text part when only a text
part is not followed by other text parts, instead of
eagerly creating a new.