Implement constrained function calling (#121)
Function calling configuration allows specifying that the model must, or
must not, call a function on the current response. By default the model
is allowed to choose whether to call a function or reply in natural
language.

Add a `FunctionCallingConfig` class and a `functionCallingConfig`
argument to the `GenerativeModel` constructor and the methods for
generating content. The classes are fields match the API model directly.

Add support for overriding the `tools` for a specific generate content
call.