Mark some flutter files as generated
Add a `.gitattributes` file with some patterns for files created by the
flutter templates.

This list is the patterns that look the least likely to be manually
edited, although there are many other files with generated (and
unedited) boilerplate that look like it _might_ change in a manual
commit at some point which will still add noise in new app PRs.
- `.metada`, `flutter/CMakeLists.txtx`,
  `GenerateDPluginRegistrant.swift`, and the `generated_*` pattern files
  have comments indicating they are generated and should not be edited.
- The `Runner.xcodeproj` and `Runner.xcworkspace` directories contain
  xml files and look very unlikely to be hand-edited.