Add a ClientController test utility
With `StubClient` the responsibilities for checking arguments and
stubbing return values are overloaded to an `EqualityMap` checking the
full content of the Uri and JSON payload and using it as a key to choose
a stubbed response. This is verbose; details are repeated in every test
that are not relevant to the test. It also has hard to diagnose
failures; when an argument is incorrect it surfaces as a missing key
with no hint about what value in the payload is wrong.

Add a `ClientController` class which more directly addresses stubbing
and argument checking individually. When only stubbing is necessary the
`verifyRequest` argument can be omitted, and any request gets the
stubbed response. Expectations checked against the request URI and JSON
payload can be as shallow or deep as necessary for the given test.

Add an `arbitraryGenerateContentResponse` variable to fill in for
responses where the parsed output isn't tested.