Unseal the Part class
The vertex AI SDK will need new subtypes for parts which are not used on
the Google AI API. For instance in vertex referring to uploaded files
includes both the Uri and a mimetype, while on Google AI an uploaded
file stores it's own mime type and the Part only includes the Uri.

Prepare to publish to keep vertex unblocked.