Include full GenerateContentRequest for countTokens (#169)
This will make the counted tokens aligned with what the model will count
if the same request were made to `generateContent`.

Extract the JSON conversion to a separate method and call it from
`generateContent`, `generateContentStream`, and `countTokens`.

Add `model` to the general `GenerateContentRequest`. It is required
when using that pattern for counting tokens, and harmless otherwise.