Use updated gemini model names (#173)
Gemini documentation is standardizing on using versioned model names.
Use `gemini-1.5-flash-latest` for most use cases. Since this model is
multimodal it can replace both `1.0-pro` and `1.0-pro-vision` usage.
Combine the models in the flutter app to one.

Use `gemini-1.5.-pro-latest` for the example with force function calling
(`FunctionCallingMode.any`) since that is unsupported in the flash
model.

Also use the pro model for the other function calling sample. The flash
model should support this in theory, but when I run it in practice it
fails to call the functions.