Remove some workarounds for the 3.0 SDK (#203)
The min SDK is now 3.1.0.

Use an `if case` with a variable assignment pattern over a separate
local variable.

Use a `??` to handle a nullable private field. The min SDK still does
not support private field promotion, but the `??` null defaulting is
slightly more readable anyway.

Remove some ignores after a fix to the `unused_result` diagnostic.