Add support for code execution (#202)
Add `CodeExecutionResult` and `ExecutableCode` which are new `Part`
subtypes that the model may reply with when code execution is enabled.
Add `CodeExecution` configuration class as a marker to enable code
execution. In the future additional configuration parameters may be
added

Add the standardized samples for code execution through the model and a
chat.

Add tests for forwarding configuration and parsing.

Bump version and prepare to publish.